refactor(cli): retain error vocabulary boundary
